Move api fetch related functions to fetch.js
